+++ backend/controllers/reviewController.js
@@ -1,11 +1,23 @@
 const reviewService = require("../services/reviewService");
-
+const movieService = require("../services/movieService");
 // Get all reviews for a movie
 const getAllReviewsForMovie = async (req, res) => {
   try {
     const { movieTitle } = req.params;
+
+    //validation for req.params
+    if (!movieTitle) {
+      throw new Error("Movie title required");
+    }
+
+    //check if the movie with provided title exists?
+    const existingMovie = await movieService.getMovieByTitle(movieTitle);
+
+    if (existingMovie === null) {
+      throw new Error("Movie doesnt exist");
+    }
     const reviews = await reviewService.getAllReviewsForMovie(movieTitle);
-    if (!reviews) {
+    if (reviews.length === 0) {
       return res
         .status(404)
         .json({ message: "No reviews found for this movie" });
@@ -13,7 +25,7 @@ const getAllReviewsForMovie = async (req, res) => {
     res.json(reviews);
   } catch (error) {
     console.error("Error getting reviews for movie:", error);
-    res.status(500).json({ message: "Internal Server Error" });
+    res.status(500).json({ message: error.message });
   }
 };
 
@@ -21,11 +33,20 @@ const getAllReviewsForMovie = async (req, res) => {
 const getAllReviewsForUserForMovie = async (req, res) => {
   try {
     const { userName, movieTitle } = req.params;
+
+    //validation for req.params
+    if (!userName) {
+      throw new Error("User name required");
+    }
+    if (!movieTitle) {
+      throw new Error("Movie Title required");
+    }
+
     const reviews = await reviewService.getAllReviewsForUserForMovie(
       userName,
       movieTitle
     );
-    if (!reviews) {
+    if (reviews.length === 0) {
       return res
         .status(404)
         .json({ message: "No reviews found for this user for this movie" });
@@ -41,6 +62,11 @@ const getAllReviewsForUserForMovie = async (req, res) => {
 const getAllReviewsForUser = async (req, res) => {
   try {
     const { userName } = req.params;
+
+    //validation for req.parms
+    if (!userName) {
+      throw new Error("User name required");
+    }
     const reviews = await reviewService.getAllReviewsForUser(userName);
     if (!reviews) {
       return res
@@ -50,7 +76,7 @@ const getAllReviewsForUser = async (req, res) => {
     res.json(reviews);
   } catch (error) {
     console.error("Error getting reviews for user:", error);
-    res.status(500).json({ message: "Internal Server Error" });
+    res.status(500).json({ message: error.message });
   }
 };
 
@@ -59,6 +85,15 @@ const updateReview = async (req, res) => {
   try {
     const { reviewId } = req.params;
     const { reviewText } = req.body;
+
+    //validation for req.body and params
+    if (!reviewId) {
+      throw new Error("Review ID required");
+    }
+    if (!reviewText) {
+      throw new Error("Updated Review text required");
+    }
+
     const updatedReview = await reviewService.updateReview(
       reviewId,
       reviewText
@@ -74,23 +109,46 @@ const updateReview = async (req, res) => {
 const deleteReview = async (req, res) => {
   try {
     const { reviewId } = req.params;
+
+    //validation for params
+    if (!reviewId) {
+      throw new Error("Review ID required");
+    }
     const deletedReview = await reviewService.deleteReview(reviewId);
     res.json(deletedReview);
   } catch (error) {
     console.error("Error deleting review:", error);
-    res.status(500).json({ message: "Internal Server Error" });
+    res.status(500).json({ message: error.message });
   }
 };
 
 //add a review
 const addReview = async (req, res) => {
   try {
+    //validation for req.body
     const reviewData = req.body;
+    if (
+      !reviewData.reviewText ||
+      !reviewData.movieTitle ||
+      !reviewData.userName
+    ) {
+      throw new Error("Missing required fields");
+    }
+
+    //check if the movie with provided title exists?
+    const existingMovie = await movieService.getMovieByTitle(
+      reviewData.movieTitle
+    );
+
+    if (existingMovie === null) {
+      throw new Error("Cannot add reviews for a movie that don't exist");
+    }
+
     const newReview = await reviewService.addReview(reviewData);
     res.status(201).json(newReview);
   } catch (error) {
     console.error("Error adding review:", error);
-    res.status(500).json({ message: "Internal Server Error" });
+    res.status(500).json({ message: error.message });
   }
 };
